[ref: hwp-349] Squash
Summary

Squashing all commits into one large commit to keep the git log clean.

[ref: hwp-349] Implement hot reload fix

Makes changes to the package.json in the form of added scripts and creates local env variables in the authDuck file.

- Scripts in the package.json
	* start-mac
	* start-pc
- Addition of local env variables for the authDuck

[ref: hwp-349] Footer style adjust and github link addition

Changes the height and padding of the footer as well as adds the link to the github repo.

- Adds inline padding to the footer to match the header
- Increases the height of the footer
- Adds link element to the footer which leads to the github repo

[ref: hwp-349] Side menu icon position

Side menu icon positioning changes from 'right: 5px' to 'right: 5%'

- Side menu icon positioning changes from 'right: 5px' to 'right: 5%'

[ref: hwp-349] Footer dividers

- Adds material UI dividers between footer links
	* These aren't showing up

[ref: hwp-349] Profile Info

- Changed some of the typography variants